pyproject.toml
[project]
name = "intelligent-customer-service"
version = "0.1.0"
description = "Enterprise-level intelligent customer service system based on LangChain + RAG"
authors = [
    {name = "dumplingsup",email = "421730156@qq.com"}
]
requires-python = "^3.10"
dependencies = [
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "chromadb (>=1.5.2,<2.0.0)",
    "openai (>=2.26.0,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "streamlit (>=1.55.0,<2.0.0)",
    "pypdf (>=6.7.5,<7.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "langgraph (>=1.0.10,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[tool.poetry.scripts]
start-api = "src.api.routes:main"
start-frontend = "streamlit.cli:main run src/frontend/app.py"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
